fix setup

+++ setup.sh
@@ -13,12 +13,15 @@ then
 fi
 if [ ! -f .env ]
 then
-  rm database.ini;
+  
   export "$(cat .env | xargs)";
 fi
-# FIXME
-echo "THIS PART IS NOT WORKING"
-echo "user     = \"$MYSQL_USER\"" > database.ini
-echo "password = \"$MYSQL_PASSWORD\"" >> database.ini
-echo "dbname   = \"$MYSQL_DATABASE\"" >> database.ini
-echo "host     = \"omeka-db\"" >> database.ini
+if [ -f .env ]
+then
+  export $(cat .env | sed 's/#.*//g' | xargs)
+  rm database.ini;
+  echo "user     = \"$MYSQL_USER\"" > database.ini
+  echo "password = \"$MYSQL_PASSWORD\"" >> database.ini
+  echo "dbname   = \"$MYSQL_DATABASE\"" >> database.ini
+  echo "host     = \"omeka-db\"" >> database.ini
+fi
